Разделяет код на отдельные модули;
