+++ js/main.js
@@ -1,136 +1,4 @@
-const NAMES = [
-  'Андрей',
-  'Мария',
-  'Петр',
-  'Кирилл',
-  'Алиса',
-  'Юлия',
-  'Валентина',
-  'Катерина',
-];
+import './util.js';
+import {createPhotoDescription} from './data.js';
 
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-const DESCRIPTION = 'Какое-то красноречивое описание для фотографии';
-
-const PICTURES = {
-  MIN: 1,
-  MAX: 25,
-};
-
-const LIKE = {
-  MIN: 15,
-  MAX: 200,
-};
-
-const AVATAR = {
-  MIN: 1,
-  MAX: 6,
-};
-
-const COMMENTS = {
-  MIN: 1,
-  MAX: 5,
-};
-
-const RANDOM_OBJECTS = 25; // 25 сгенерированных объектов;
-const RANDOM_MESSAGES = {
-  MIN: 1,
-  MAX: 2,
-};
-
-const ID = {
-  MIN: 1,
-  MAX: 100,
-};
-
-function getRandomInt(min, max) {
-  if (isNaN(Number(min)) || isNaN(Number(max))) {
-    return NaN;
-  }
-
-  if (min < 0 || max < 0) {
-    return NaN;
-  }
-
-  min = Math.ceil(min);
-  max = Math.floor(max);
-
-  if (min === max) {
-    return NaN;
-  }
-
-  if (min > max) {
-    const temp = min;
-    min = max;
-    max = temp;
-  }
-
-  return Math.floor(Math.random() * (max - min + 1) + min);
-}
-
-getRandomInt(1, 10);
-
-//Функция для проверки максимальной длины строки
-function checkStringLength(string, maxLength) {
-  return string.length <= maxLength;
-}
-
-checkStringLength('Meow', 150);
-
-//Функции от Кекса, возвращающая случайное целое число из переданного диапазона включительно
-function getRandomPositiveInteger (a, b) {
-  if (a < 0 || b < 0) {
-    return NaN;
-  }
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-}
-
-const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
-
-const generateMessage = () => getRandomArrayElement(MESSAGES);
-
-const generateComments = () => {
-  const commentId = getRandomPositiveInteger (ID.MIN, ID.MAX);
-  const avatarText = `img/avatar-${getRandomPositiveInteger(AVATAR.MIN, AVATAR.MAX)}.svg`;
-  const messageText = Array.from({length: getRandomPositiveInteger (RANDOM_MESSAGES.MIN, RANDOM_MESSAGES.MAX)}, generateMessage);
-  const userName = getRandomArrayElement(NAMES);
-
-  return {
-    id: commentId,
-    avatar: avatarText,
-    message: messageText,
-    name: userName,
-  };
-};
-
-const generatePhotoDesc = () => {
-  const photoId = getRandomPositiveInteger (PICTURES.MIN, PICTURES.MAX);
-  const urlText = `photos/${getRandomPositiveInteger(PICTURES.MIN, PICTURES.MAX)}.jpg`;
-  const likesValue = getRandomPositiveInteger (LIKE.MIN, LIKE.MAX);
-  const totalComments = Array.from({length: getRandomPositiveInteger (COMMENTS.MIN, COMMENTS.MAX)}, generateComments);
-
-  return {
-    id: photoId,
-    url: urlText,
-    description: DESCRIPTION,
-    likes: likesValue,
-    comments: totalComments,
-  };
-};
-
-const photoDescription = Array.from({length:RANDOM_OBJECTS}, generatePhotoDesc);
-
-const runArray = () => photoDescription;
-
-runArray();
+createPhotoDescription();
